feat: refactor mobile navbar and add block
